+++ src/models/SubStoreModel/index.ts
@@ -1,0 +1,1 @@
+export { default as SubStoreModel } from './SubStoreModel';