+++ src/stores/RootStore/index.ts
@@ -1,0 +1,3 @@
+export { default as RootStore } from './RootStore';
+
+export * from './types';