+++ src/stores/UserStore/index.ts
@@ -1,0 +1,3 @@
+export { default as UserStore } from './UserStore';
+
+export * from './types';