+++ types/typings.d.ts
@@ -1,5 +1,6 @@
 import { WindowType } from '../src';
 
 declare global {
+  // eslint-disable-next-line @typescript-eslint/no-empty-interface
   interface Window extends WindowType {}
 }